+++ components/form-response-base/question-element/question-multiple-choice-input.vue
@@ -1,10 +1,10 @@
 <script lang="ts" setup>
-import type { ResponseSettingsMultipleChoice } from '~/components/based-response-form/based-response-form-settings';
+import type { ResponseSettingsMultipleChoice } from '~/components/form-response-base/form-response-base-settings';
 import type {
   ResponseFormQuestionInvalid,
   ResponseFormQuestionSettingsBase,
   ResponseFormQuestionSettingsMultipleChoice,
-} from '~/components/based-response-form/questionnaire-settings';
+} from '~/components/form-response-base/questionnaire-settings';
 
 type ModelValue = ResponseSettingsMultipleChoice &
   ResponseFormQuestionSettingsMultipleChoice &
@@ -72,3 +72,4 @@ const name = computed(() => `multiple-choice-input-${createId()}`);
   padding-left: 8px;
 }
 </style>
+~/components/form-response-base/questionnaire-settings
